updating data.py, resp. what is now called xydata.py
